+++ app/src/main/java/org/dhis2/usescases/teiDashboard/dashboardfragments/teidata/TEIDataFragment.kt
@@ -632,7 +632,7 @@ class TEIDataFragment : FragmentGlobalAbstract(), TEIDataContracts.View {
         guid: String,
         orgUnitUid: String,
         trackedEntityInstanceId: String,
-        ageInMonths: Long,
+        ageInMonths: Double,
     ) {
         val biometricsClient = get(requireContext())
         val extras: HashMap<String, String> = HashMap()
@@ -643,7 +643,7 @@ class TEIDataFragment : FragmentGlobalAbstract(), TEIDataContracts.View {
     override fun registerBiometrics(
         orgUnitUid: String,
         trackedEntityInstanceUId: String,
-        ageInMonths: Long
+        ageInMonths: Double
     ) {
         val biometricsClient = get(requireContext())
         val extras: HashMap<String, String> = HashMap()
